correción mapeo dato id entidad clase

+++ backend/src/main/java/GymBook/backend/entities/Inscripcion.java
@@ -32,9 +32,13 @@ public class Inscripcion {
         this.fechaInscripcion = fechaInscripcion;
     }
 
-    // Getters y Setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 
     public Usuario getUsuario() { return usuario; }
     public void setUsuario(Usuario usuario) { this.usuario = usuario; }
